+++ tests/models/test_models.py
@@ -6,6 +6,11 @@ from pathlib import Path
 from pydantic import ValidationError
 
 from BALSAMIC.models.cache_models import ReferenceMeta, ReferenceUrlsModel
+from BALSAMIC.models.metrics_model import (
+    MetricModel,
+    MetricValidationModel,
+    MetricConditionModel,
+)
 from BALSAMIC.models.models import (
     VCFAttributes,
     VarCallerFilter,
@@ -19,9 +24,6 @@ from BALSAMIC.models.models import (
     UMIParamsTNscope,
     ParamsVardict,
     ParamsVEP,
-    MetricModel,
-    MetricConditionModel,
-    MetricValidationModel,
     AnalysisPonModel,
 )
